requirements.txt
gradio
mysql-connector-python
pandas
python-dotenv
